Updated identity.joomla.org AddThis scripts
Despite the creation of the AddThis script on identity.joomla.org, we still got issues on the property itself. 

The cookiescript, is validating the ' $siteConfig->scripts->addthis->false ' value as true, because the if statement checks for !== 'undefined' values. Therefore, the addthis script is appended to body with false values.

This commit, sets the ' $siteConfig->scripts->addthis = true ' and the ' $siteConfig->scripts->addthisId = <addthis id> '