Push script functions and cookie initialization
All the necessary scripts, have been moved into seperate functions. Each function, assigns a code part into head or body, according to each services purpose. In order to push these functions into the necessary spot, the property<Script> variable's value and type is checked for not being undefined (for type) and 'undefined' (for value). This means that in template.php, the values should be either 'undefined' for not initialized scripts, or the id's value.

Finally, the cookies initialization has been set in separate arrays in order to make the CookieControl.delete('<cookie>') function, dynamic for every cookie in the array( each array = a cookie category).