AddThis rename and value changes among scripts
The `$siteConfig->scripts->addthis` variable changed to `$siteConfig->scripts->carbonads`. Besides that, all script values have been changed to `'undefined'`. As it is mentioned on PR #209, all scripts that don't have an id value or a flag value, should be set as undefined. If an id/flag value exists, then it should be replace the undefined value.