Accessibility and IAB TCF implementation
The accessibility and IAB TCF features have been enabled. A remaining issue must be solved which includes the manipulation of the `propertyGoogleSynd` variable. This variable must enable or disable the `googlesyndication` script which will fetch ads from the Global Vendor List (as mention in IAB's TCF).

At this right moment, the case study of the `googlesyndication` ads is CJO property.